REFACTOR
Reorganization of code architecture
