+++ src/main/java/com/rebels/alliance/domains/Traitor.java
@@ -1,4 +1,4 @@
-package com.rebels.alliance.domains.models;
+package com.rebels.alliance.domains;
 
 //public class Traitor extends Rebel {
